Fetch GitHub Access Token from SSM
This means we only ever need AWS_ variables available to complete the
install, making developer UX quite a bit nicer.

The downside is that we need to handle dev-account access, which means
knowing the cross-account Arn, which we don't want to commit in this
public repository.

On balance, this is acceptable because:

- Local usage is not negatively impacted (they always use prod)
- CI is minimally-impacted