Include setup-python in ourselves
When this Action was first created, you could not use other Actions via
uses in composite definitions. This meant we asked our users to
setup-python before using this, so we could install cfn-flip.

That limitation is gone, so we can call it ourselves now :tada:

In the future, we will probably gate this on a check to see if Python is
already adequately configured, so we don't over-configure in a Workflow
where there needs to be a different and specific Python setup. But we'll
defer that for now.