Export SLACK_ variables
As described in the new README section, this makes it less cumbersome to
use the `rtCamp` action directly. This was the only other feature of the
`platform-deploy-action` that we're trying to obviate.

Before, one action did setup/deploy/notify:

```yaml
- uses: platform-deploy-action
  with:
    platform-setup-token: x
    platform-version: y
    environment: z
    slack-url: w
```

Now, we have more composable pieces:

```yaml
- uses: setup-platform-action
  with:
    token: x
    version: y
    environment: z

- run: platform deploy --tag '...'

- uses: rtCamp/action-slack-notify
  with:
    SLACK_WEBHOOK: w
```

This is only _slightly_ more work for the end-user, but there's way less
stuff hiding behind the scenes. It's more transparent and flexible this
way. That's the hope anyway.