Cache according to resource input (#14)
When this action is used in a matrix, only one of the jobs will successfully write to the cache. This can be observed in `freckle/progress`, where runs of `progress (db)` and `progress (api)` show the same cache being restored, but the latter job unable to write to the cache. This is because the `job_id` is constant across matrix runs.

Although we'd like to consolidate this deploy into a single job, I think this is unexpected and can be addressed by using the `resource` input in the cache key. In that repo, future runs would get distinct caches for the `api` and `db` resources. This should also preserve caches where this action is used without a `resource` key, using the `restore-keys` function of the cache action.