Default to latest versions if un-specified
Every time we release a new version we have to come and update these
defaults. It was assumed at the time that we'd want to let the new
versions "bake" for a while first, so the manual step would be valuable,
but that is not the case. We simply do the busy work automatically every
time (or forget). Given that, let's just make it the automatic behavior
to lookup the latest release if not explicitly configured.