Rename github-private-key input
I made a naming mistake. I thought I new better, but I did not.

I thought, when I saw `app-id` as the official action's input, that was
`(app-)id`. It is the "Id" of an "App" after all, right?

So then, when I then saw their `private-key` input, I assumed they had a
historical accident, and it should be `(app-)private-key`. I decided to
"fix" this, by using `app-private-key` and `APP_PRIVATE_KEY` in any of
the names I controlled, including this input.

I now understand, through other documentation, actions, and various
naming conventions, that (for better or worse), GitHub's intent there is
actually `[github-]app-id`: it's an "`AppId`", not an "`Id` for an
`App`". I feel like this choice will forever cause problems, but it
seems to be what they did. And, of course, they are (so far)
internally-consistent with that intent by using `[github-]private-key`,
which is what I should've matched.

This PR begins the add-transition-remove cycle by adding
`github-private-key` here.